backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
reportlab==4.2.0
requests==2.31.0
python-dotenv==1.0.1
pydantic==2.7.1
python-multipart==0.0.9
psycopg2-binary==2.9.9
alembic==1.13.1
